refactor(classify): decompose llm.py into focused submodules (REFACTOR-001, #267)
Split the 1150-LOC `creek/classify/llm.py` into a package of focused
modules so prompt construction, provider clients, response parsing,
the FEAT-017 confidence bias, batch concurrency, and the orchestrator
each live behind their own boundary:

- `prompts.py` — CLASSIFICATION_PROMPT, sanitisation, prompt builder.
- `providers.py` — AnthropicProvider plus Ollama call / health-check
  helpers.
- `parsing.py` — YAML/reasoning split, enum parsers, validate_response,
  frequency/voice apply helpers.
- `calibration.py` — FEAT-017 confidence-threshold bias for Mode /
  Orientation / Dosage plus the wavelength apply helper that uses it.
- `batch.py` — BatchStats and the thread-pool driver.
- `orchestrator.py` — LLMClassifier wires it all together; thin
  delegation methods (`_build_prompt`, `_call_ollama`,
  `validate_response`, `_apply_classification`) preserve the existing
  test surface.
- `__init__.py` — re-export shim keeps every `from creek.classify.llm
  import …` site (and `creek.classify.llm.{httpx,time}.*` patches)
  working.

Pure refactor: no signature or behaviour changes. All 3851 tests still
pass unchanged, mypy strict and ruff are clean, complexity stays at A
average / B max, aggregate coverage is 93.75%.
